refactor(tests): split test_cloud into per-source mirrors
tests/tools/cloud/test_cloud.py (526) covered three source modules at
once. Split to mirror tools/cloud/, no coverage change (39 tests
redistributed, none dropped):

  test_aws.py      <- TestCheckS3Buckets (aws.py)
  test_azure.py    <- TestCheckAzureBlobContainers / TestCheckAzureSasTokens
                      (azure.py)
  test_services.py <- TestCheckUnauthenticatedDatabases /
                      TestCheckSensitiveFiles / TestCheckAdminPanels /
                      TestGranularPanels (services.py)

test_databases.py already mirrors the cloud/databases subpackage. Largest
resulting file is test_services.py at 381 lines, under the 450 bar.
